Type safety in OCLEngine

+++ include/common/oclengine.hpp
@@ -379,7 +379,7 @@ public:
         if (dev < -1) {
             throw std::runtime_error("Invalid device selection: " + std::to_string(dev));
         }
-        int lDev = (dev == -1) ? GetDefaultDeviceID() : dev;
+        size_t lDev = (dev == -1) ? GetDefaultDeviceID() : dev;
 
         if (size == 0) {
             return activeAllocSizes[lDev];
@@ -395,7 +395,7 @@ public:
     }
     void ResetActiveAllocSize(const int64_t& dev)
     {
-        int lDev = (dev == -1) ? GetDefaultDeviceID() : dev;
+        size_t lDev = (dev == -1) ? GetDefaultDeviceID() : dev;
         std::lock_guard<std::mutex> lock(allocMutex);
         // User code should catch std::bad_alloc and reset:
         activeAllocSizes[lDev] = 0;
